+++ app/Controllers/Profil.php
@@ -14,10 +14,21 @@ class Profil extends BaseController
         $this->profilsModel = model('Profil');
     }
 
+    private function isAuthorized(): bool
+    {
+        $user = auth()->user();
+        return $user->inGroup('admin') || $user->inGroup('rhu');
+    }
+
     //------------------------------------------
     //Liste
-    public function liste(): string
+    public function liste()
     {
+        if (!$this->isAuthorized()) {
+            return redirect('accueil');
+        }
+
+
         $listeProfils = $this->profilsModel->findall();
         return view(
             'profils_liste',
@@ -32,12 +43,20 @@ class Profil extends BaseController
 
     public function ajout()
     {
+        if (!$this->isAuthorized()) {
+            return redirect('accueil');
+        }
+
         return view('profils_ajoute');
     }
 
     //Create
     public function create()
     {
+        if (!$this->isAuthorized()) {
+            return redirect('accueil');
+        }
+
         $profilsData = $this->request->getpost();
         $this->profilsModel->save($profilsData);
 
@@ -47,6 +66,10 @@ class Profil extends BaseController
     // Modifier
     public function modif($profil)
     {
+        if (!$this->isAuthorized()) {
+            return redirect('accueil');
+        }
+
         $modifProfils = $this->profilsModel->find($profil);
 
         return view(
@@ -60,6 +83,10 @@ class Profil extends BaseController
     // Update
     public function update()
     {
+        if (!$this->isAuthorized()) {
+            return redirect('accueil');
+        }
+
         $profilsData = $this->request->getpost();
         $this->profilsModel->save($profilsData);
 
@@ -69,6 +96,10 @@ class Profil extends BaseController
     // Delete
     public function delete()
     {
+        if (!$this->isAuthorized()) {
+            return redirect('accueil');
+        }
+
         $profilsData = $this->request->getpost();
         $this->profilsModel->delete($profilsData['ID_PROFIL']);
